In Scenario 7, offer a way for the player to kill the enemy leader without angering the orcs
If you spared Graktar and found the cowardly orc archer, then the archer
may make the killing blow without angering Graktar; but at the cost of
leaving the party. If you insist on keeping the archer, Graktar is
angered anyway.

If you did not spare Graktar, his son is not so reasonable; this chance
is not offered.